remove postgres adapter unwrap on convert to params
We only accept connection params now. It might be possible to convert
between the errors "properly", but right now there seems to be some
fragmentation between the shared and tokio crates that makes this
difficult.

For the moment, I'm ok with leaving as is since its not too hard to
convert into the params if you really only have a string, as seen in the tests.