Fix deadlock when getting new connection
@enzious reported that they had issues getting new connections after
some startup time. Testing revealed that there was a deadlock in the
Pool::connection method caused by holding a mutex lock for the entire
function body. This change ensures that the mutex lock is held only as
long as is required. I also added a test that verifies the deadlock is
no longer present. This test passes with this code change, and fails
without it.