Check is_valid when pulling from pool
The is_valid method should be used when pulling a connection from the
pool to ensure that we do not hand an invalid connection to the caller.
If an invalid connection is pulled off the pool, it will be removed and
a new connection will be spawned onto the pool in its place.

I also added more channels to synchronize the status of the connection
wrappers with the async driver tasks.