breaking: Allow test-on-checkout to be optional
This is a breaking change because our previous config object (struct
with all public fields) does not allow for config options to be added
without making a breaking change. Since a break was required anyways,
I felt it prudent to change the config in such a way that will allow us
to add options to the config in the future without a breaking change.