Add the ability to configure connection timeouts
Previously, we did not directly support connection timeouts, relying on
the end-user to call timeout on the returned future. It will be easier
for users to configure a default timeout that will be used for all
connection attempts. The default behavior is unchanged, if you do not
configure a timeout, there will be no timeout.