connect_no_timeout: Add comments for the reason of the loop
We want to purge all the invalid connections, e.g. due to the
network failure and recreate all the connections fresh in a
single connection request, so that we can get the connection
in the next attempt.  And also, once all those invalid
connections were purged, the next connect_no_timeout just block
to create a new connection, as there is no valid connections
in the pool.