queue: Use crossbeam_queue::ArrayQueue for idle connections pool
Since the pool size is limitted to the max connections, we can
use crossbeam_queue::ArrayQueue to maintain the idle connections,
which is a bit faster than unbound crossbewam_queue::SegQueue.

The size of the queue is set to twice the size of the max connections,
just to avoid the unnecessary error happens during the connection
push back, even though, technically, it shouldn't happen.  It exposes
the error counter when it happens, so that we can observe the behavior
as well.