test: conn-pool exhaustion reproduction
There is a problem with failing over in playerd that is suspected to be
a connection pool exhaustion bug. In production, we noticed that
`playerd` or `onepush` would sometimes get into a bad state during a
failover. This diff adds a reproduction case:

1. `cd l337-redis`
2. `docker-compose up -d`
3. `cargo test --package l337-redis --lib -- tests::it_works_on_pause
--exact --nocapture`
4. Wait for a few seconds, you should see pool statistics being printed
and an "Attempting to ping now..." message printed as well
5. Kill the underlying redis with `docker kill`
6. Hang